Fix voting on webhook triggered builds (again)
Conditional using "client_payload" in "env" section resulted
in either boolean (True) if payload was present or empty string
if not. When "true", there was no way to extract change and patchset number
in the "report" stage of the build.

Change variable name and set it to be either true/false instead of
string.

Signed-off-by: Karol Latecki <karol.latecki@nutanix.com>