build_qcow: cache the VM image for faster spin up of every job
Artifact download takes between 1-3 minutes on every job in
the per-patch matrix. Instead we can cache the image reducing,
download to 20-30 seconds.

Currently only the one and only VM image is cached.
It takes almost 3GB, while cache is limited to 10GB.
While we could explore caching other artifacts shared on
per-patch basis - like SPDK repo, but we need to be careful
not to boot the VM image out of the cache.

Note: used hashFiles() instead of checking cache-hit output,
because it was not set despite hitting the cache.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>