infra: add webhook forwarder script code
Add script for forwarding webhooks between Gerrit instance
and Github Actions dispatcher. This comes with an example
docker-compose.yml file which resembles current Gerrit
configuration and can be used for local development.

As the README.md describes - this is not a copy of the
working production configuration! Forwarder code matches
what's actually used, but other configuration files are
to be considered more as a reference.

Signed-off-by: Karol Latecki <karol.latecki@nutanix.com>