Scan and un-verify outdated Gerrit changes
This workflow will query Gerrit and look for changes which
were not updated for a while. If there was no activity for:
* 2 weeks - the script will suggest user to rebase their change
* 4 weeks - the script will set Verified -1 vote on change,
            and ask user to rebase.

Workflow runs daily as 12AM UTC, or can be run on-demand
by priviledged users.

There's also an additional 12 weeks cut-off date. Changes
older than that will NOT be commented on to avoid polluting
Gerrit dashboards with very old changes (e.g. 2-3 years old).

Signed-off-by: Karol Latecki <karol.latecki@nutanix.com>