per-patch: post SPDK patch details in summary very early
As part of summary job we put the patch details into the
artifacts and GitHub Summary. Yet that happens at the very
end of the job.

While we have not checked out the SPDK just yet,
we can post details we got from the JSON from Gerrit.
Removed the Gerrit link in report summary as to not duplicate
too much.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>