per-patch: account for NVMe-oF RDMA tests results
Looks like we only counted the common job results
in the summary job.

This was probably missed because autorun_post job previously
could have failed with no RDMA tests. If the set of tests in
TCP and RDMA are the same, it thinks all tests were executed.
They were, just not in both variants as we would expect.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>