Add vfio_user tests
Add spdk/test/vfio_user to common tests.
Because GH runner is a VM, the tests are run
in a container to avoid three layers of
virtualization.

Neither spdk/autotest.sh or spdk/autorun.sh are used
because spdk/setup.sh (used along the way) in it's
current state fails due to lack of uio_pci_generic and
vfio-pci drivers. Instead, for now, build and tests are
executed using hard-coded commands.

Signed-off-by: Karol Latecki <karol.latecki@nutanix.com>